+++ src/es/eternalmangas/build.gradle
@@ -1,7 +1,9 @@
 ext {
     extName = 'EternalMangas'
     extClass = '.EternalMangas'
-    extVersionCode = 1
+    themePkg = 'mangaesp'
+    baseUrl = 'https://eternalmangas.com'
+    overrideVersionCode = 0
     isNsfw = true
 }
